+++ src/client/src/containers/PageAzureLogin/index.tsx
@@ -18,7 +18,7 @@ import { IOption } from "../../types/option";
 import { azureMessages } from "../../mockData/azureServiceOptions";
 import { AppState } from "../../reducers";
 import AzureSubscriptions from "../AzureSubscriptions";
-import AzureStudent from "../AzureStudent";
+import AzureStudent from "./AzureStudent";
 import Title from "../../components/Title";
 import RootAction from "../../actions/ActionType";
 import keyUpHandler from "../../utils/keyUpHandler";
