+++ src/client/src/containers/PageDetails/Details/styles.module.css
