+++ src/client/src/containers/PageSelectFrameworks/FrameworkCard/DependencyInfo/messages.ts
