+++ src/client/src/containers/PageSelectFrameworks/FrameworkCard/DependencyInfo/styles.module.css
