+++ src/client/src/containers/RightSidebar/ServicesList/CosmosDBSelection/styles.module.css
