+++ src/client/src/containers/RightSidebar/ServicesList/index.tsx
@@ -1,13 +1,13 @@
 import { connect } from "react-redux";
-import { AppState } from "../../reducers";
-import { getServicesSelector } from "../../selectors/servicesSelector";
+import { AppState } from "../../../reducers";
+import { getServicesSelector } from "../../../selectors/servicesSelector";
 import React from "react";
 import { injectIntl, InjectedIntlProps } from "react-intl";
 import messages from "./strings";
 import styles from "./styles.module.css";
 import CosmosDBSelection from "./CosmosDBSelection";
-import AppServiceSelection from "../AppServiceSelection";
-import { ServiceState } from "../../reducers/wizardSelectionReducers/services";
+import AppServiceSelection from "../../AppServiceSelection";
+import { ServiceState } from "../../../reducers/wizardSelectionReducers/services";
 import _ from "lodash";
 
 interface IProps {
